+++ aequilibrae/parameters.py
@@ -1,6 +1,7 @@
 import os
 import yaml
-
+from warnings import warn
+from copy import deepcopy
 
 class Parameters:
     """
@@ -31,21 +32,30 @@ class Parameters:
 
     def __init__(self):
         """ Loads parameters from file. The place is always the same. The root of the package"""
-        path = os.path.dirname(os.path.realpath(__file__))
-        self.file = os.path.join(path, "parameters.yml")
-        with open(self.file, "r") as yml:
-            self.parameters = yaml.load(yml, Loader=yaml.SafeLoader)
+
+        proj_path = os.environ.get('AEQUILIBRAE_PROJECT_PATH', '')
+        default_path = os.path.dirname(os.path.realpath(__file__))
+
+        self.file = os.path.join(proj_path, "parameters.yml")
+        self.file_default = os.path.join(default_path, "parameters.yml")
+
+        with open(self.file_default, "r") as yml:
+            self._default = yaml.load(yml, Loader=yaml.SafeLoader)
+
+        if os.path.isfile(self.file):
+            with open(self.file, "r") as yml:
+                self.parameters = yaml.load(yml, Loader=yaml.SafeLoader)
+        else:
+            warn('No pre-existing parameter file exists for this project. Will use default')
+            self.parameters = deepcopy(self._default)
+
 
     def write_back(self):
         """Writes the parameters back to file"""
-        stream = open(self.file, "w")
-        yaml.dump(self.parameters, stream, default_flow_style=False)
-        stream.close()
+        with open(self.file, "w") as stream:
+            yaml.dump(self.parameters, stream, default_flow_style=False)
 
     def restore_default(self):
         """Restores parameters to generic default"""
-        path = os.path.dirname(os.path.realpath(__file__))
-        default_file = os.path.join(path, "parameter_default.yml")
-        with open(default_file, "r") as yml:
-            self.parameters = yaml.load(yml, Loader=yaml.SafeLoader)
+        self.parameters = self._default
         self.write_back()
