+++ aequilibrae/paths/results/skim_results.py
@@ -15,7 +15,7 @@ class SkimResults:
           from aequilibrae.paths.results import SkimResults
 
           proj = Project()
-          proj.load('path/to/project.sqlite')
+          proj.load('path/to/project/folder')
           proj.network.build_graphs()
           # Mode c is car in this project
           car_graph = proj.network.graphs['c']
