+++ docs/source/network.rst
@@ -189,7 +189,7 @@ you can import the network without creating indices, as shown below.
   from aequilibrae.project import Project
 
   p = Project()
-  p.new('path/to/project/file.sqlite')
+  p.new('path/to/project/new/folder')
   p.network.create_from_osm(place_name='my favorite place')
   p.conn.close()
 
@@ -212,7 +212,7 @@ with the option for creating such indices.
   from aequilibrae.project import Project
 
   p = Project()
-  p.new('path/to/project/file.sqlite')
+  p.new('path/to/project/new/folder/')
   p.network.create_from_osm(place_name='my favorite place', spatial_index=True)
   p.conn.close()
